add read structure test to structure
