+++ tests/qchem/hf_tests/test_ref_files/gdb3.mol5.XYZ
